Use semaphores instead of libUnique.
Now we don't need the libUnique package as the dependency.

When the first instance starts, it acquires a semaphore, and
(well, this is a bit hacky) stores its PID in the semaphore value.

The second instance checks, if semaphore is present, gets the PID of the
first instance, and terminates it.