Fix semaphore not being deleted
When running, closing and running the program again, it won't run due to
the semaphore not being deleted upon normal program exit; it stays in
the kernel, which causes the secondly ran instance to think there is
still the first one. It tries to kill it, then exits.

Fixes issue 6.