Rename LICENSE → COPYING
LICENSE file wasn't included into distribution package, and COPYING file is.
That is the default behaviour of autotools. And since I don't know how this
stuff works and made all the configuration scripts by cut-and-paste, I don't
know how to change the behaviour. And I'm too lazy to research. This is
standard anyway.

Fixes #3.