Add ability to execute external viewer
To do so one should doubleclick the date one's interested in, provided the
    external_command = some command line with "%Y-%m-%d" strftime substitutions
option is given in the configuration file.

Documentation will be updated shortly.

P.S. I thought I don't want to do command line parsing and all, so this job is
delegated to the shell. Currently it is hardcoded "/bin/sh", but I hope it will
do for all the users. If you have some trouble, please file a bug.

Fixes #4.