Fix current executable path obtaining on BSD
That is, we used ftok with the path to current executable which we obtained
from /proc. But /proc has different layouts on Linux and on BSD.

On Linux we have:

    /proc/{getpid()}/exe

On BSD we have:

    /proc/{getpid()}/file

Additionally, we can use neat shortcuts to getpid() call, which are:

On Linux:

    /proc/self/exe

On BSD:

    /proc/curproc/file

Which we can use since we do conditional compilation already.

Fixes #13.