Added close_on_mouseleave option
With this option, the window will close only when the mouse leaves the
calendar window.
There is a similar option `close_on_unfocus` which closes the window
when it loses focus. The problem is that the window gains focus when it
gets created and when the user first drags the mouse pointer over other
windows before reaching the calendar window, the calendar closes.
This happens when the window manager is set to automatically focus
windows when the mouse pointer hovers over them.